fixed runtime errors, changed auto moving from PID to a ramping profile

+++ src/edu/wpi/first/wpilibj/templates/subsystems/SpeedControllerPID.java
@@ -44,7 +44,7 @@ public class SpeedControllerPID extends PIDSubsystem {
         driveMotors[1].set(speed);
     }
 
-    public double getD() {
+    public int getD() {
         return -enc.get() * direction;
     }
